Removing implicit conversions@

+++ engine/rasterizer_integer.cpp
@@ -146,9 +146,9 @@ void LeRasterizer::rasterList(LeTriList * trilist)
 			float ubot = tri->us[vb] / tri->zs[vb];
 			float vtop = tri->vs[vt] / tri->zs[vt];
 			float vbot = tri->vs[vb] / tri->zs[vb];
-			float d = cmax(fabs(utop - ubot), fabs(vtop - vbot));
+			float d = cmax(fabsf(utop - ubot), fabsf(vtop - vbot));
 
-			int r = (d * bmp->ty + dy * 0.5f) / dy;
+			int r = (int) ((d * bmp->ty + dy * 0.5f) / dy);
 			int l = LeGlobal::log2i32(r);
 			l = cmin(l, bmp->mmLevels - 1);
 			bmp = bmp->mipmaps[l];
